pyproject.toml
[tool.poetry]
name = "Stochastic-Pricing-Model"
version = "0.0.0"
description = "Prototype for Stochastic Pricing Model"

package-mode = false
authors = [
    "Mirco Di Gioia <mirco.di.gioia@outlook.com>"
]

repository = "https://github.com/Ocrim93/Stochastic-Pricing-Model/tree/main"
homepage = "https://github.com/Ocrim93/Stochastic-Pricing-Model/tree/main"

# README file(s) are used as the package description
readme = ["README.md"]

# Keywords (translated to tags on the package index)
keywords = ["packaging", "poetry"]

[tool.poetry.dependencies]
# Compatible Python versions
python="^3.10"
# Standard dependency with semver constraints
aiohttp = "^3.8.1"
# Dependency with extras
requests = { version = "^2.28", extras = ["security"] }
# Version-specific dependencies with prereleases allowed
tomli = { version = "^2.0.1", python = "<3.11", allow-prereleases = true }
# Git dependencies
cleo = { git = "https://github.com/python-poetry/cleo.git", branch = "main" }
# Optional dependencies (installed by extras)
pendulum = { version = "^2.1.2", optional = true }

# Dependency groups are supported for organizing your dependencies
pandas = "^2.3.1"
numpy = "^1"
yfinance = "^0.2.65"
scipy = "^1"
jupyter = "^1.1.1"
ipykernel = "^7.0.1"
loguru = "^0.7.3"
pyyaml = "^6.0.3"
[tool.poetry.group.dev.dependencies]
pytest = "^7.1.2"
pytest-cov = "^3.0"

# ...and can be installed only when explicitly requested
[tool.poetry.group.docs]
optional = true
[tool.poetry.group.docs.dependencies]
Sphinx = "^5.1.1"

# Python-style entrypoints and scripts are easily expressed
[tool.poetry.scripts]
my-script = "my_package:main"